requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
Pillow==10.4.0